+++ CHANGES.md
@@ -388,3 +388,6 @@
 - Sonic: added back for the update server is actually still working, at least for now
 - Portable release: update Python for Win10x64 to v3.14.5
 - Portable release: update Python modules
+
+### Pre-release (work-in-progress)
+- Sonic: backup removed due to be unreliable